[#10] Split factories onto separate files
