+++ app/services/factory/howl_sound_factory.ts
@@ -1,0 +1,7 @@
+export class HowlSoundFactory {
+    public static createSound(soundPath: string): Howl {
+        return new Howl({
+            src: [soundPath]
+        });
+    }
+}
